Revert "Fix health bars"
This reverts commit 4b1a99c06df5b2c259ec7e258bea8f3b69c4f4ed.

+++ Earth-FAE/Assets/Scripts/enemyBear.cs
@@ -624,7 +624,7 @@ public class enemyBear : MonoBehaviour, Enemy
         for(int i = 0; i < d && health > 0; i++)
         {
             health--;
-            healthBar.fillAmount = (float) health / maxHealth;
+            healthBar.fillAmount = health / maxHealth;
         }
         Debug.Log(health);
         if (health == 0)
